Separating pre & post Sign Up comps
