+++ src/app/auth/page.module.css
@@ -48,6 +48,8 @@
 
 .submitBtn {
   background: var(--cherryPie);
+  margin-top: 1.5rem;
+  margin-bottom: 1rem;
 }
 
 .callToAction {
